LayoutTests/fast/mediastream/mediastreamtrack-video-white-balance-mode.html

<!DOCTYPE html>
<html>
<head>
    <meta charset='utf-8'>
    <title>Camera track whiteBalanceMode</title>
    <script src='../../resources/testharness.js'></script>
    <script src='../../resources/testharnessreport.js'></script>
</head>
<body>
    <canvas id='canvas'></canvas>
    <video controls autoplay width=640px height=480px playsInline id='video'></video>
    <script>

        let stream;
        promise_test(async (test) => {
            stream = await navigator.mediaDevices.getUserMedia({ video: { width : 640 } });
            assert_equals(stream.getVideoTracks()[0].getSettings().whiteBalanceMode, undefined, 'settings-1');
            assert_equals(stream.getVideoTracks()[0].getCapabilities().whiteBalanceMode, undefined, 'capabilities-1');

            stream = await navigator.mediaDevices.getUserMedia({ video: { width: 640, whiteBalanceMode : 'single-shot' } });
            assert_equals(stream.getVideoTracks()[0].getSettings().whiteBalanceMode, 'single-shot', 'settings-2');

            await stream.getVideoTracks()[0].applyConstraints({ whiteBalanceMode : 'manual' });
            assert_equals(stream.getVideoTracks()[0].getSettings().whiteBalanceMode, 'manual', 'settings -3');

        }, 'getUserMedia whiteBalanceMode');
    </script>
</body>
</html>
